NLP Based Text Summarization for Kannada
Text Summarization is a Natural Language Processing (NLP) technique used to automatically generate a shorter and meaningful version of a long document while preserving the key information, important ideas, and overall context. The main goal of text summarization is to reduce the content length without losing important points, making it easier for users to quickly understand large amounts of information.